refactor(ui): move a soon-to-be commonly used constant to shareable location
